adds check image move condition.

+++ app/Http/Controllers/ImageUploadController.php
@@ -5,11 +5,13 @@ namespace App\Http\Controllers;
 use App\Models\Image;
 use App\Models\TemporaryFile;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Log;
 
 class ImageUploadController extends Controller
 {
     public function list(){
         $imageList = Image::latest()->get();
+//        dd($imageList);
         return view('image_upload.list', compact('imageList'));
     }
 
@@ -22,10 +24,18 @@ class ImageUploadController extends Controller
             'image' => $request->image,
         ]);
         $temporaryFile = TemporaryFile::where('folder', $request->image)->first();
+//        dd($temporaryFile);
         if ($temporaryFile){
-            $imageStore->addMedia(storage_path('app/public/images/tmp'. $request->image. '/'. $temporaryFile->filename))->toMediaCollection();
-            rmdir(storage_path('app/public/images/tmp'. $request->image));
-            $temporaryFile->delete();
+            $filePath = storage_path('app/public/images/tmp/' . $request->image . '/' . $temporaryFile->filename);
+            $media = $imageStore->addMedia($filePath)->toMediaCollection();
+
+            if ($media) {
+                rmdir(storage_path('app/public/images/tmp/' . $request->image));
+                $temporaryFile->delete();
+            } else {
+                // Log or handle failure case, if the media could not be added
+                Log::error('Failed to add media for file: ' . $temporaryFile->filename);
+            }
         }
         return redirect()->route('image.list')->with('success', "Image Stored Successfully.");
     }
